Extract TypeDeclarationMap from flow and typescript to a common utils.js
