+++ packages/react-native-codegen/src/parsers/flow/components/componentsUtils.js
@@ -11,9 +11,9 @@
 'use strict';
 
 import type {ASTNode} from '../utils';
-import type {TypeDeclarationMap} from '../utils.js';
 import type {NamedShape} from '../../../CodegenSchema.js';
 const {getValueFromTypes} = require('../utils.js');
+import type {TypeDeclarationMap} from '../../utils';
 
 function getProperties(
   typeName: string,
