+++ packages/react-native-codegen/src/parsers/flow/components/index.js
@@ -9,7 +9,7 @@
  */
 
 'use strict';
-import type {TypeDeclarationMap} from '../utils';
+import type {TypeDeclarationMap} from '../../utils';
 import type {CommandOptions} from './options';
 import type {ComponentSchemaBuilderConfig} from './schema.js';
