+++ packages/react-native-codegen/src/parsers/typescript/components/commands.js
@@ -14,7 +14,7 @@ import type {
   NamedShape,
   CommandTypeAnnotation,
 } from '../../../CodegenSchema.js';
-import type {TypeDeclarationMap} from '../utils.js';
+import type {TypeDeclarationMap} from '../../utils';
 const {parseTopLevelType} = require('../parseTopLevelType');
 
 type EventTypeAST = Object;
