+++ packages/react-native-codegen/src/parsers/typescript/modules/index.js
@@ -22,8 +22,7 @@ import type {
   Nullable,
 } from '../../../CodegenSchema.js';
 
-import type {TypeDeclarationMap} from '../utils.js';
-import type {ParserErrorCapturer} from '../../utils';
+import type {ParserErrorCapturer, TypeDeclarationMap} from '../../utils';
 import type {NativeModuleTypeAnnotation} from '../../../CodegenSchema.js';
 
 const {
